+++ simplified-app/src/main/java/org/nypl/simplified/app/reader/ReaderSettings.java
@@ -105,7 +105,8 @@ public final class ReaderSettings implements ReaderSettingsType
   @Override public void setFontScale(
     final float s)
   {
-    final Double x = Double.valueOf(Math.max(50.0, Math.min(s, 200.0)));
+    final Double x =
+      Double.valueOf(Math.max(50.0, Math.min((double) s, 200.0)));
     ReaderSettings.LOG.debug("font size: {}%", x);
 
     final Editor e = this.settings.edit();
